molecule_app/requirements.txt
absl-py==1.3.0
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1724167852130/work
aiohttp==3.8.4
aiosignal==1.3.1
ale-py==0.8.1
altgraph==0.17.3
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1716290248287/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1717693030552/work
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
baselines==0.1.5
beautifulsoup4==4.11.1
bleach==5.0.1
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1715091184126/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
cachetools==5.2.0
cchardet @ file:///home/conda/feedstock_root/build_artifacts/cchardet_1695670604369/work
certifi==2023.5.7
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
cloudpickle==2.2.1
cmake==3.27.2
colorama==0.3.9
contourpy==1.0.5
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1721521267415/work
cx-Freeze==6.15.2
cycler==0.11.0
Cython==3.0.2
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1723662581198/work
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.7
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1704321475663/work
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing==1.1.1
fairscale==0.4.13
fastjsonschema==2.16.2
filelock==3.12.3
fire==0.5.0
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1712667726126/work
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1715865800631/work
gast==0.4.0
glfw==2.5.9
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.50.0
gym-notices==0.0.8
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
h5py==3.8.0
hpack==4.0.0
httpcore @ file:///home/conda/feedstock_root/build_artifacts/httpcore_1711596990900/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1708530890843/work
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1724106245647/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna==3.4
imageio==2.31.1
importlib-resources==5.12.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1724187233579/work
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1713372668944/work
jax==0.4.12
jedi==0.18.1
Jinja2==3.1.2
jiter @ file:///home/conda/feedstock_root/build_artifacts/jiter_1723011410442/work
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.6
llvmlite==0.40.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
ml-dtypes==0.2.0
mpi4py==3.1.4
mpmath==1.3.0
mujoco==2.2.0
multidict==6.0.4
multiprocess @ file:///home/conda/feedstock_root/build_artifacts/multiprocess_1706514640841/work
natsort==5.1.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook==6.5.1
notebook_shim==0.2.0
numba==0.57.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808312889/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai @ file:///home/conda/feedstock_root/build_artifacts/openai_1724215031650/work
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
opencv-python==4.6.0.66
opt-einsum==3.3.0
overrides==7.4.0
packaging==21.3
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1715897614105/work
pandocfilters==1.5.0
parso==0.8.3
patchelf==0.17.2.1
pickleshare==0.7.5
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1719903544255/work
pipenv==2023.9.8
platformdirs==3.10.0
progressbar2==4.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==4.23.2
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix @ file:///home/conda/feedstock_root/build_artifacts/pyarrow-hotfix_1700596371886/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1720293063581/work
pydantic_core @ file:///home/conda/feedstock_root/build_artifacts/pydantic-core_1720041207964/work
pygame==2.1.0
Pygments==2.13.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyOpenGL==3.1.7
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python-utils==3.6.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1723018227415/work
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors @ file:///home/conda/feedstock_root/build_artifacts/safetensors_1722937815382/work
seaborn==0.12.1
Send2Trash==1.8.0
sentencepiece==0.1.99
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.3.2.post1
stable-baselines==2.10.0
stack-data==0.5.1
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.0
terminaltables==3.1.0
tiktoken==0.3.3
tinycss2==1.2.1
tokenizers @ file:///home/conda/feedstock_root/build_artifacts/tokenizers_1713402723682/work/bindings/python
torch==2.0.1
torch-tb-profiler==0.4.0
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1724214724759/work
# Editable install with no version control (tree-of-thoughts-llm==0.1.0)
-e /home/potatosalad/Documents/projects/tree-of-thought-llm
triton==2.0.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2023.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1719391292974/work
virtualenv==20.24.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1724051738774/work
yacs==0.1.8
yarl==1.9.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1723591248676/work
zmq==0.0.0
zstandard==0.23.0
